+++ R/guide-.R
@@ -50,7 +50,7 @@ new_guide <- function(..., available_aes = "any", super) {
 
   # Validate theme settings
   if (!is.null(params$theme)) {
-    check_object(params$theme, is.theme, what = "a {.cls theme} object")
+    check_object(params$theme, is_theme, what = "a {.cls theme} object")
     validate_theme(params$theme, call = caller_env())
     params$direction <- params$direction %||% params$theme$legend.direction
   }
@@ -65,6 +65,10 @@ new_guide <- function(..., available_aes = "any", super) {
     available_aes = available_aes
   )
 }
+
+#' @export
+#' @rdname is_tests
+is_guide <- function(x) inherits(x, "Guide")
 
 #' @section Guides:
 #'
@@ -377,10 +381,10 @@ Guide <- ggproto(
   # Renders tickmarks
   build_ticks = function(key, elements, params, position = params$position,
                          length = elements$ticks_length) {
-    if (!inherits(elements, "element")) {
+    if (!is_element(elements)) {
       elements <- elements$ticks
     }
-    if (!inherits(elements, "element_line")) {
+    if (!is_element(elements, "line")) {
       return(zeroGrob())
     }