+++ R/margins.R
@@ -8,7 +8,10 @@ margin <- function(t = 0, r = 0, b = 0, l = 0, unit = "pt") {
   class(u) <- c("margin", class(u))
   u
 }
-is.margin <- function(x) {
+
+#' @export
+#' @rdname is_tests
+is_margin <- function(x) {
   inherits(x, "margin")
 }