+++ R/plot-construction.R
@@ -51,9 +51,9 @@
   # can be displayed in error messages
   e2name <- deparse(substitute(e2))
 
-  if      (is.theme(e1))  add_theme(e1, e2, e2name)
-  else if (is.ggplot(e1)) add_ggplot(e1, e2, e2name)
-  else if (is.ggproto(e1)) {
+  if      (is_theme(e1))  add_theme(e1, e2, e2name)
+  else if (is_ggplot(e1)) add_ggplot(e1, e2, e2name)
+  else if (is_ggproto(e1)) {
     cli::cli_abort(c(
       "Cannot add {.cls ggproto} objects together.",
       "i" = "Did you forget to add this object to a {.cls ggplot} object?"