+++ R/theme-elements.R
@@ -21,6 +21,8 @@
 #'   a blank element among its parents will cause this element to be blank as
 #'   well. If `FALSE` any blank parent element will be ignored when
 #'   calculating final element state.
+#' @param type For testing elements: the type of element to expect. One of
+#'   `"blank"`, `"rect"`, `"line"` or `"text"`.
 #' @return An S3 object of class `element`, `rel`, or `margin`.
 #' @examples
 #' plot <- ggplot(mpg, aes(displ, hwy)) + geom_point()
@@ -144,6 +146,22 @@ element_text <- function(family = NULL, face = NULL, colour = NULL,
   )
 }
 
+#' @export
+#' @rdname element
+is_element <- function(x, type = "any") {
+  switch(
+    type %||% "any",
+    any   = inherits(x, "element"),
+    rect  = inherits(x, "element_rect"),
+    line  = inherits(x, "element_line"),
+    text  = inherits(x, "element_text"),
+    blank = inherits(x, "element_blank"),
+    # TODO: ideally we accept more elements from extensions. We need to
+    # consider how this will work with S7 classes, where ggplot2 doesn't know
+    # about the extension's class objects.
+    FALSE
+  )
+}
 
 #' @export
 print.element <- function(x, ...) utils::str(x)